code: introduce net_fifo struct and move mac out of nethandler
struct net_fifo is a container for network-related info required to
properly send data via the timedc_avtp container. This is inteded to be
a per-project file used at compile-time that defines all netfifos.

Work has also started to move dst out of nethandler and into timedc_avtp
as the address should be a multicast address.