code: drop size from pdu_update()
We set the size when creating the DU, so this is stored inside for later
usage. It is reasonable to expect the data to be of stable size, so we
place this as a requirement.

Either way, we expect to have a pipe/fifo/lvhcan to read from, and then
it will use the size from the DU to determine how many bytes to read and
detect errors.