code: add TX DataUnit, create listener in other end of pipe
When we use a channel over the network, we provide a pipe to the TimedC
program. In the other end of this pipe is a waiting thread that reads
the expected number of bytes from the pipe, updates the dataunit and
sends it via the correct socket.

This somewhat convoluted step is done to abstract away the transport
mechanism and allows the TimedC program to treat the Netchan as any
other fifo.