code: use avtpdu_cshdr as struct in callbacks
The callback and feed_pdu (wrongly) used the timedc_avtp container as
argument, but the data we get from the network will be the AVTP Data
Unit. Update to use this and move size into private header so that
read()/write() knows the correct size to use.

This also updates the relevant tests (most of them, the change is fairly
complex).