code: drop ifname and hmap_sz from net_fifo.h and add as cli-arg
Having a static requirement for which ethernet port to use is not
productive. We want to lock the net_fifo struct as this will be common
across the distributed system, but local nic should not be locked to a
binary.

This affects quite a large portion of the code, so cleanup where NIC is
used.