code: add READ/WRITE macro to timedc_avtp
This is the final abstraction to read and write data to a channel. It
will get channel specifics from net_fifo_chans and construct the sockets
from there.

In essence, to send a payload:

	NETFIFO_TX(mcast42);
	WRITE(mcast42, &payload_tx);

And correspondingly on the receiver:

	NETFIFO_RX(mcast42);
	READ(mcast42, &payload_rx);

mcast is the label given in .name in net_fifo_chans.