code: grab timestamp from PTP device
CLOCK_TAI yields time in TAI format without leap-seconds, but is not
synched to PTP unless phc2sys also runs. Instead, since we are working
under the assumption that a ptp daemon is keeping the NIC we're using in
sync, grab the timestamp directly from there.

This adds helpers to handle this.