code: add READ_WAIT and WRITE_WAIT, class specific waits
When using (READ|WRITE)_WAIT, caller will wait for 2 (or 50) ms and
continue at the same time the other end of the pipe. This is a fairly
straightforward way to use timedc_netfifo for a rendezvous.

Current implementation uses relative delay, this will be changed to
absolute to reduce error in the future.