code: use absolute time for sleeping and disable cstate
To reduce/avoid drift, use absolute time for wakeup as it simplifies the
logic a lot.

By disabling cstate, we reduce the wakeup-error from ~200us to ~1.4us

This gives the READ_DELAY() a cross-system accuracy of:
 gPTP_error_A + wakeup_A + gPTP_error_B + wakeup_B

gPTP error is in the order of 4-20ns RMS, the wakeup error is
~1.3-1.5us, giving us a safe estimate of the 2 machines waking up less
than 4 us apart.