remove unstable and/or pointless tests
The MRPD-tests are highly unstable and prone to non-deterministic input
from the network. Most of the tests focues on the third-party library
mrpd to format strings correctly and is not really a core part of
netchan.

The macro-tests are outdated and we are moving away from macros, so drop
these tests and reduce false failures.