Split out Tx-channel create into own function
A lot of the logic for creating Rx and Tx channels were kept in
chan_create_standalone(), which was supposed to be an easy-to-use
function for the macros (i.e. you don't have to create the nethandler
first).

By moving all Tx-related logic into own function, we can start placing
the appropriate steps in chan_create(), chan_create_tx() and
chan_create_rx() instead of keepign everything in _standalone().