Do not store channel-name.
The channel struct kept this as "iface name", which it was not used as,
but rather the name of the channel used in a manifest.

As the manifest is a helper and not realy relevant for the channel,
using the name for anything but finding the right entry in the array is
pointless.

Drop the name and stop testing for it.