Rename net_fifo struct to 'channel_attrs'
net_fifo was never a good name as it was used to list the attributes of
a given channel. The name was also a leftover from when channels was
called net_fifo.

Consequently, arrays of channel attributes in manifest are renamed to
"nc_channels" and macros and helpers updated to this.