Use raw rx-sockets to open properly in promiscous mode
For distributed systems spread over multiple machines, but where a
subset of the nodes run on the /same/ host, multicast back to same host
would not work as the kernel will filter them out.

Since the socket was opened for TSN-packets and not RAW, promiscous-mode
did not work.

Instead of only opening lo in promiscous, open Rx in promiscous
regardless and filter out TSN-types instead.

This will add extra overhead as /every/ single incoming frame on the
interface will be forwarded, but it is the only way to ensure we
consistently receive locally generated multicast frames.