Do not sleep outside periodic_timer
netchan will look at previous ts and determine when next tx should
occur. netchan will not block sender but instead expects it to properly
pace the tx-rate (alternatively, use wait_for_tx_slot()).

In manychan we use the periodic_timer, which should handle this. The
problem arises when we between instantiating pt and using it, throw in a
usleep().

Even though it can be useful to sleep to ensure that collaborative
threads are properly sync'd, this throws a spanner into the tx-slots
used, so to avoid unexplainable delays or Tx-queue overflows, do not use
usleep() (or other means of delaying) except periodic timer.