Add a timeout to await_listener
If talker starts before listener, we run the risk of the listener never
receiving the advertise message. Likewise, the talker would block
forever waiting for a listener.

This adds a timeout counter which allows netchan_srp() to re-announce
the stream, ensuring that a new listener will be able to connect.

Note that the talker will *not* re-announce the stream once a listener
has attached, nor does it properly handle the case when a listener leaves.