Make channel setup asynchronous and guard internal state
Setting up a channel can take time when we use SRP as both talker and
listener will block until the other end has arrived.

This change forks out channel setup when SRP is enabled, allowing the
application to continue whilst the initial talker-listener setup is
performed in the background.

Correspondingly, a ready and timedwait is added to allow the application
to wait for a channel to become ready, but at its own leisure.