Move label after variable declaration
In commit d9ce590111e9 (Add a timeout to await_listener) a goto
statement was added when mrp_await_listener() timed out. This would
ensure that whilst waiting for a listener, the talker would announce the
stream regularly.

The label should have been placed after the variable declaration to
avoid throwing "error: a label can only be part of a statement and a
declaration is not a statement".