reduce logging when tx fails with deadline miss
If we are sending frames to the NIC with SO_TXTIME out of order, we run
the risk of not being able to send one or more frames at the specified
instant.

This is especially relevant when multiple processes each tie into the
NIC using SO_TXTIME (such as when we run several instances on the same
host in a simulated distributed setup).

Apart from adding a qdisc to re-sort (which requires longer deadlines
for frames to send), this problem should not be as painful on systems
with only a single application using netchan.