Introduce a soft-stop option to channel and nethandler
Improve the stop-action of the system to ensure a clean shutdown and
proper SRP tear-down regardless of where in a startup sequence a client
is.

This introduces nh_stop() and extra support functionality. By calling
nh_stop(), the system is not torn down, but all streams are stopped and
if SRP is in use, unannounced/leave stream events are triggered.

It should not be the responsibility of a client to know the exact
sequence of functions to call, an overreaching nh_stop() should take
care of all internal contingencies.