Start adding TAS and CBS as two stream classes
It has been shown that SO_TXTIME and the Qdisc machinery is not capable
of keeeping pace with streams with a high packet rate. Once the rate
exceeds 500Hz, standard linux distributions have a hard time enusring
consistent (and timely) Qdisc dequeue  behavior.

We need to differentiate between the two types (low-rate, but accurate
txtime and high-rate, constant bandwidth) and map each to two different
hw-queues. This is the first step to re-introduce CBS whilst also
keeping the SO_TXTIME(TAS) implementation.