Add timestamp logging to _cbs send and use tai-clock for all timestamps
We use the timestamp from the NIC for HW tx and Rx, and if we don't use
CLOCK_TAI for these timestamps, then some will be 37 seconds off.

Ideally, CBS should use CLOCK_REALTIME for timestamps, but when merging
and analyzing the logs, this makes things messy.

A potential problem will be _cbs_send_at() as this compares the
timestamp with CLOCK_TAI, whereas it should probably be
CLOCK_REALTIME. This should be handled, ideally all clocks should just
be CLOCK TAI (we expect PTP to be running after all)